+++ lib/ts_api_handler.dart
@@ -6,6 +6,7 @@ export 'package:dio/dio.dart' show CancelToken, Response;
 
 export 'src/api_cache_helper.dart' show ApiCacheHelper;
 export 'src/api_service.dart' show ApiServices;
+export 'src/utils/api_log_options.dart' show ApiLogOptions;
 export 'src/utils/error_source_extension.dart' show ErrorSource;
 export 'src/utils/failure.dart' show Failure;
 export 'src/utils/response_code.dart' show ResponseCode;